Moved CartTestSuite and UserTestSuite from ava/com/kodilla/ecommercee to ava/com/kodilla/ecommercee/domain
