+++ src/test/java/com/kodilla/ecommercee/domain/UserTestSuite.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.domain;
 
 import com.kodilla.ecommercee.dbServices.UserDbService;
 import com.kodilla.ecommercee.domain.Order;
